Migration to JUnit 5 - avoid using AbstractMojoTestCase
it finishes migration

also:
- refactor MavenProjectResourcesStub - remove not needed additional stubs and not used methods

- remove not used test resources
