+++ src/test/java/org/apache/maven/plugins/resources/ResourcesMojoTest.java
@@ -18,42 +18,66 @@
  */
 package org.apache.maven.plugins.resources;
 
+import javax.inject.Inject;
+
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileReader;
 import java.io.IOException;
-import java.util.Collections;
 import java.util.LinkedList;
-import java.util.List;
 import java.util.Properties;
 
+import org.apache.maven.api.di.Provides;
+import org.apache.maven.api.plugin.testing.InjectMojo;
+import org.apache.maven.api.plugin.testing.MojoParameter;
+import org.apache.maven.api.plugin.testing.MojoTest;
 import org.apache.maven.execution.DefaultMavenExecutionRequest;
 import org.apache.maven.execution.MavenExecutionRequest;
 import org.apache.maven.execution.MavenSession;
-import org.apache.maven.model.Resource;
-import org.apache.maven.plugin.testing.AbstractMojoTestCase;
 import org.apache.maven.plugins.resources.stub.MavenProjectResourcesStub;
+import org.apache.maven.project.MavenProject;
 import org.codehaus.plexus.util.FileUtils;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
 
-public class ResourcesMojoTest extends AbstractMojoTestCase {
-    private final String defaultPomFilePath = "/target/test-classes/unit/resources-test/plugin-config.xml";
+import static org.apache.maven.api.plugin.testing.MojoExtension.setVariableValueToObject;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+@MojoTest
+class ResourcesMojoTest {
+
+    @TempDir
+    private File testRootDir;
+
+    @Inject
+    private MavenProject projectStub;
+
+    @Inject
+    private MavenSession mavenSession;
+
+    @Provides
+    @SuppressWarnings("unused")
+    private MavenProject projectStub() {
+        return new MavenProjectResourcesStub(testRootDir.getAbsolutePath());
+    }
 
     /**
      * test mojo lookup, test harness should be working fine
      */
-    public void testHarnessEnvironment() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-
+    @Test
+    @InjectMojo(goal = "resources")
+    void testHarnessEnvironment(ResourcesMojo mojo) {
         assertNotNull(mojo);
     }
 
-    public void testResourceDirectoryStructure() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourceDirectoryStructure");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testResourceDirectoryStructure(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -64,12 +88,6 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.addFile("notpackage/test/file2.txt");
         project.setupBuildEnvironment();
 
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -81,15 +99,14 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertTrue(FileUtils.fileExists(resourcesDir + "/notpackage/test"));
     }
 
-    public void testResourceDirectoryStructureRelativePath() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourceDirectoryStructure_RelativePath");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    @MojoParameter(name = "outputDirectory", value = "../relative_dir")
+    void testResourceDirectoryStructureRelativePath(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
-        project.setOutputDirectory("../relative_dir");
         project.addFile("file4.txt");
         project.addFile("package/file3.nottest");
         project.addFile("notpackage/file1.include");
@@ -97,15 +114,9 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.addFile("notpackage/test/file2.txt");
         project.setupBuildEnvironment();
 
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
         mojo.execute();
 
-        String resourcesDir = project.getOutputDirectory();
+        String resourcesDir = mojo.getOutputDirectory().getAbsolutePath();
 
         assertTrue(FileUtils.fileExists(resourcesDir + "/file4.txt"));
         assertTrue(FileUtils.fileExists(resourcesDir + "/package/file3.nottest"));
@@ -114,11 +125,11 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertTrue(FileUtils.fileExists(resourcesDir + "/notpackage/test"));
     }
 
-    public void testResourceEncoding() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("encoding");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    @MojoParameter(name = "encoding", value = "UTF-8")
+    void testResourceEncoding(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -126,13 +137,6 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.setResourceFiltering(0, true);
         project.setupBuildEnvironment();
 
-        setVariableValueToObject(mojo, "encoding", "UTF-8");
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -140,11 +144,10 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertTrue(FileUtils.fileExists(resourcesDir + "/file4.txt"));
     }
 
-    public void testResourceInclude() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourceInclude");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testResourceInclude(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -171,12 +174,6 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.addInclude("**/test/file*");
         project.addInclude("**/package/*.include");
 
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -188,11 +185,10 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertFalse(FileUtils.fileExists(resourcesDir + "/notpackage/nottest/file.txt"));
     }
 
-    public void testResourceExclude() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourceExclude");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testResourceExclude(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -220,12 +216,6 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.addExclude("**/notpackage*");
         project.addExclude("**/notpackage*/**");
 
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -237,14 +227,10 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertFalse(FileUtils.fileExists(resourcesDir + "/notpackage/nottest/file.txt"));
     }
 
-    /**
-     * @throws Exception
-     */
-    public void testResourceTargetPath() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourceTargetPath");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testResourceTargetPath(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -257,12 +243,6 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.addFile("notpackage/test/file2.txt");
         project.setupBuildEnvironment();
 
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -274,14 +254,10 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertTrue(FileUtils.fileExists(resourcesDir + "/org/apache/maven/plugin/test/notpackage/test"));
     }
 
-    /**
-     * @throws Exception
-     */
-    public void testResourceSystemPropertiesFiltering() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourceSystemProperties_Filtering");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testResourceSystemPropertiesFiltering(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -289,20 +265,11 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.setResourceFiltering(0, true);
         project.setupBuildEnvironment();
 
-        // setVariableValueToObject(mojo,"encoding","UTF-8");
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
-        setVariableValueToObject(mojo, "escapeWindowsPaths", Boolean.TRUE);
-
         MavenExecutionRequest request = new DefaultMavenExecutionRequest();
         request.setSystemProperties(System.getProperties());
 
-        MavenSession mavenSession = new MavenSession(null, null, request, null);
-        setVariableValueToObject(mojo, "session", mavenSession);
+        mavenSession.getSystemProperties().putAll(System.getProperties());
+
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -316,18 +283,14 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         }
         File fileFromFiltering = new File(props.getProperty("current-working-directory"));
 
-        assertTrue(fileFromFiltering.getAbsolutePath() + " does not exist.", fileFromFiltering.exists());
+        assertTrue(fileFromFiltering.exists(), fileFromFiltering.getAbsolutePath() + " does not exist.");
         assertEquals(userDir.getAbsolutePath(), fileFromFiltering.getAbsolutePath());
     }
 
-    /**
-     * @throws Exception
-     */
-    public void testResourceProjectPropertiesFiltering() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourceProjectProperties_Filtering");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testResourceProjectPropertiesFiltering(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -336,13 +299,6 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.addProperty("user.dir", "FPJ kami!!!");
         project.setupBuildEnvironment();
 
-        // setVariableValueToObject(mojo,"encoding","UTF-8");
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -351,15 +307,10 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertContent(resourcesDir + "/file4.txt", checkString);
     }
 
-    /**
-     * @throws Exception
-     */
-    public void testProjectPropertyFilteringPropertyDestination() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project =
-                new MavenProjectResourcesStub("resourcePojectProperty_Filtering_PropertyDestination");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testProjectPropertyFilteringPropertyDestination(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -370,14 +321,6 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         // setup dummy property
         project.setDescription("c:\\\\org\\apache\\test");
 
-        // setVariableValueToObject(mojo,"encoding","UTF-8");
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
-        setVariableValueToObject(mojo, "escapeWindowsPaths", Boolean.TRUE);
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -386,14 +329,10 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertContent(resourcesDir + "/file4.properties", checkString);
     }
 
-    /**
-     * @throws Exception
-     */
-    public void testPropertyFilesFiltering() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourcePropertyFiles_Filtering");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testPropertyFilesFiltering(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
         LinkedList<String> filterList = new LinkedList<>();
 
         assertNotNull(mojo);
@@ -404,13 +343,8 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.setupBuildEnvironment();
         filterList.add(project.getResourcesDirectory() + "filter.properties");
 
-        // setVariableValueToObject(mojo,"encoding","UTF-8");
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
         setVariableValueToObject(mojo, "buildFilters", filterList);
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
+
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -419,14 +353,10 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertContent(resourcesDir + "/file4.properties", checkString);
     }
 
-    /**
-     * @throws Exception
-     */
-    public void testPropertyFilesExtra() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourcePropertyFiles_Extra");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testPropertyFilesExtra(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
         LinkedList<String> filterList = new LinkedList<>();
 
         assertNotNull(mojo);
@@ -437,13 +367,8 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.setupBuildEnvironment();
         filterList.add(project.getResourcesDirectory() + "filter.properties");
 
-        // setVariableValueToObject(mojo,"encoding","UTF-8");
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
         setVariableValueToObject(mojo, "filters", filterList);
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
+
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -452,14 +377,10 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertContent(resourcesDir + "/extra.properties", checkString);
     }
 
-    /**
-     * @throws Exception
-     */
-    public void testPropertyFilesMainAndExtra() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("resourcePropertyFiles_MainAndExtra");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testPropertyFilesMainAndExtra(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
         LinkedList<String> filterList = new LinkedList<>();
         LinkedList<String> extraFilterList = new LinkedList<>();
 
@@ -470,20 +391,14 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.addFile("extra-filter.properties", "dir2:testdir2");
         project.setResourceFiltering(0, true);
 
-        project.cleanBuildEnvironment();
         project.setupBuildEnvironment();
 
         filterList.add(project.getResourcesDirectory() + "filter.properties");
         extraFilterList.add(project.getResourcesDirectory() + "extra-filter.properties");
 
-        // setVariableValueToObject(mojo,"encoding","UTF-8");
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
         setVariableValueToObject(mojo, "buildFilters", filterList);
         setVariableValueToObject(mojo, "filters", extraFilterList);
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
+
         mojo.execute();
 
         String resourcesDir = project.getOutputDirectory();
@@ -496,15 +411,11 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
     /**
      * Validates that a Filter token containing a project property will be resolved before the Filter is applied to the
      * resources.
-     *
-     * @throws Exception
      */
-    public void testPropertyFilesFilteringTokensInFilters() throws Exception {
-        final File testPom = new File(getBasedir(), defaultPomFilePath);
-        final ResourcesMojo mojo = lookupMojo("resources", testPom);
-        final MavenProjectResourcesStub project =
-                new MavenProjectResourcesStub("resourcePropertyFiles_Filtering_TokensInFilters");
-        final List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    void testPropertyFilesFilteringTokensInFilters(ResourcesMojo mojo) throws Exception {
+        final MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
         final LinkedList<String> filterList = new LinkedList<>();
 
         assertNotNull(mojo);
@@ -516,13 +427,8 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.setupBuildEnvironment();
         filterList.add(project.getResourcesDirectory() + "filter.properties");
 
-        // setVariableValueToObject(mojo,"encoding","UTF-8");
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
         setVariableValueToObject(mojo, "buildFilters", filterList);
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
+
         mojo.execute();
         final String resourcesDir = project.getOutputDirectory();
 
@@ -531,11 +437,11 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         assertContent(resourcesDir + "/file4.properties", checkString);
     }
 
-    public void testWindowsPathEscapingDisabled() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("windows-paths");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    @MojoParameter(name = "escapeWindowsPaths", value = "false")
+    void testWindowsPathEscapingDisabled(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -545,16 +451,7 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.addFile("path-listing.txt", "base path is ${basePath}\ndocuments path is ${docsPath}");
         project.setResourceFiltering(0, true);
 
-        project.cleanBuildEnvironment();
         project.setupBuildEnvironment();
-
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
-        setVariableValueToObject(mojo, "escapeWindowsPaths", Boolean.FALSE);
 
         mojo.execute();
 
@@ -567,11 +464,11 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
                 FileUtils.fileRead(new File(resourcesDir, "path-listing.txt")));
     }
 
-    public void testWindowsPathEscapingEnabled() throws Exception {
-        File testPom = new File(getBasedir(), defaultPomFilePath);
-        ResourcesMojo mojo = lookupMojo("resources", testPom);
-        MavenProjectResourcesStub project = new MavenProjectResourcesStub("windows-paths");
-        List<Resource> resources = project.getBuild().getResources();
+    @Test
+    @InjectMojo(goal = "resources")
+    @MojoParameter(name = "escapeWindowsPaths", value = "true")
+    void testWindowsPathEscapingEnabled(ResourcesMojo mojo) throws Exception {
+        MavenProjectResourcesStub project = (MavenProjectResourcesStub) this.projectStub;
 
         assertNotNull(mojo);
 
@@ -581,17 +478,7 @@ public class ResourcesMojoTest extends AbstractMojoTestCase {
         project.addFile("path-listing.txt", "base path is ${basePath}\ndocuments path is ${docsPath}");
         project.setResourceFiltering(0, true);
 
-        project.cleanBuildEnvironment();
         project.setupBuildEnvironment();
-
-        setVariableValueToObject(mojo, "project", project);
-        setVariableValueToObject(mojo, "resources", resources);
-        setVariableValueToObject(
-                mojo, "outputDirectory", new File(project.getBuild().getOutputDirectory()));
-        setVariableValueToObject(mojo, "buildFilters", Collections.emptyList());
-        setVariableValueToObject(mojo, "useBuildFilters", Boolean.TRUE);
-
-        setVariableValueToObject(mojo, "escapeWindowsPaths", Boolean.TRUE);
 
         mojo.execute();
