+++ src/test/java/org/apache/maven/plugins/resources/stub/MavenProjectResourcesStub.java
@@ -19,17 +19,143 @@
 package org.apache.maven.plugins.resources.stub;
 
 import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Properties;
 
+import org.apache.maven.model.Build;
 import org.apache.maven.model.Resource;
+import org.apache.maven.project.MavenProject;
+import org.codehaus.plexus.util.FileUtils;
 
-public class MavenProjectResourcesStub extends MavenProjectBuildStub {
+public class MavenProjectResourcesStub extends MavenProject {
+    private String testRootDir;
 
-    private File baseDir;
+    private Properties properties;
 
-    public MavenProjectResourcesStub(String id) throws Exception {
-        super(id);
+    private String description;
+
+    private Build build;
+
+    private String srcDirectory;
+
+    private String buildDirectory;
+
+    private String outputDirectory;
+
+    private String testOutputDirectory;
+
+    private String resourcesDirectory;
+
+    private String testResourcesDirectory;
+
+    private ArrayList<String> fileList;
+
+    private ArrayList<String> directoryList;
+
+    private HashMap<String, String> dataMap;
+
+    public MavenProjectResourcesStub(String testRootDir) {
+        this.testRootDir = testRootDir;
+        properties = new Properties();
+        build = new Build();
+        fileList = new ArrayList<>();
+        directoryList = new ArrayList<>();
+        dataMap = new HashMap<>();
+
+        if (!FileUtils.fileExists(testRootDir)) {
+            FileUtils.mkdir(testRootDir);
+        }
+
+        setupBuild();
+
         setupResources();
         setupTestResources();
+    }
+
+    public void addProperty(String key, String value) {
+        properties.put(key, value);
+    }
+
+    public void addFile(String name) {
+        if (isValidPath(name)) {
+            fileList.add(name);
+        }
+    }
+
+    public void addFile(String name, String data) {
+        File fileName = new File(name);
+
+        addFile(name);
+        dataMap.put(fileName.getName(), data);
+    }
+
+    public String getOutputDirectory() {
+        return outputDirectory;
+    }
+
+    public String getTestOutputDirectory() {
+        return testOutputDirectory;
+    }
+
+    public String getResourcesDirectory() {
+        return resourcesDirectory;
+    }
+
+    /**
+     * returns true if the path is relative
+     * and false if absolute
+     * also returns false if it is relative to
+     * the parent
+     */
+    private boolean isValidPath(String path) {
+        return !path.startsWith("c:") && !path.startsWith("..") && !path.startsWith("/") && !path.startsWith("\\");
+    }
+
+    private void setupBuild() {
+        // check getBasedir method for the exact path
+        // we need to recreate the dir structure in
+        // an isolated environment
+        srcDirectory = testRootDir + "/src";
+        buildDirectory = testRootDir + "/target";
+        outputDirectory = buildDirectory + "/classes";
+        testOutputDirectory = buildDirectory + "/test-classes";
+        resourcesDirectory = srcDirectory + "/main/resources/";
+        testResourcesDirectory = srcDirectory + "/test/resources/";
+
+        build.setDirectory(buildDirectory);
+        build.setOutputDirectory(outputDirectory);
+        build.setTestOutputDirectory(testOutputDirectory);
+
+        properties.setProperty("project.build.sourceEncoding", "UTF-8");
+    }
+
+    public void setupBuildEnvironment() throws Exception {
+        // populate dummy resources and dummy test resources
+
+        // setup src dir
+        if (!FileUtils.fileExists(resourcesDirectory)) {
+            FileUtils.mkdir(resourcesDirectory);
+        }
+
+        if (!FileUtils.fileExists(testResourcesDirectory)) {
+            FileUtils.mkdir(testResourcesDirectory);
+        }
+
+        createDirectories(resourcesDirectory, testResourcesDirectory);
+        createFiles(resourcesDirectory, testResourcesDirectory);
+
+        // setup target dir
+        if (!FileUtils.fileExists(outputDirectory)) {
+            FileUtils.mkdir(outputDirectory);
+        }
+
+        if (!FileUtils.fileExists(testOutputDirectory)) {
+            FileUtils.mkdir(testOutputDirectory);
+        }
     }
 
     public void addInclude(String pattern) {
@@ -40,28 +166,8 @@ public class MavenProjectResourcesStub extends MavenProjectBuildStub {
         build.getResources().get(0).addExclude(pattern);
     }
 
-    public void addTestInclude(String pattern) {
-        build.getTestResources().get(0).addInclude(pattern);
-    }
-
-    public void addTestExclude(String pattern) {
-        build.getTestResources().get(0).addExclude(pattern);
-    }
-
     public void setTargetPath(String path) {
         build.getResources().get(0).setTargetPath(path);
-    }
-
-    public void setTestTargetPath(String path) {
-        build.getTestResources().get(0).setTargetPath(path);
-    }
-
-    public void setDirectory(String dir) {
-        build.getResources().get(0).setDirectory(dir);
-    }
-
-    public void setTestDirectory(String dir) {
-        build.getTestResources().get(0).setDirectory(dir);
     }
 
     public void setResourceFiltering(int nIndex, boolean filter) {
@@ -70,11 +176,68 @@ public class MavenProjectResourcesStub extends MavenProjectBuildStub {
         }
     }
 
+    private void createDirectories(String parent, String testparent) {
+        File currentDirectory;
+
+        for (String directory : directoryList) {
+            currentDirectory = new File(parent, "/" + directory);
+
+            if (!currentDirectory.exists()) {
+                currentDirectory.mkdirs();
+            }
+
+            // duplicate dir structure in test resources
+            currentDirectory = new File(testparent, "/" + directory);
+
+            if (!currentDirectory.exists()) {
+                currentDirectory.mkdirs();
+            }
+        }
+    }
+
+    private void createFiles(String parent, String testparent) throws IOException {
+        File currentFile;
+
+        for (String file : fileList) {
+            currentFile = new File(parent, file);
+
+            // create the necessary parent directories
+            // before we create the files
+            if (!currentFile.getParentFile().exists()) {
+                currentFile.getParentFile().mkdirs();
+            }
+
+            if (!currentFile.exists()) {
+                currentFile.createNewFile();
+                populateFile(currentFile);
+            }
+
+            // duplicate file in test resources
+            currentFile = new File(testparent, file);
+
+            if (!currentFile.getParentFile().exists()) {
+                currentFile.getParentFile().mkdirs();
+            }
+
+            if (!currentFile.exists()) {
+                currentFile.createNewFile();
+                populateFile(currentFile);
+            }
+        }
+    }
+
+    private void populateFile(File file) throws IOException {
+        String data = dataMap.get(file.getName());
+
+        if ((data != null) && file.exists()) {
+            try (FileOutputStream outputStream = new FileOutputStream(file)) {
+                outputStream.write(data.getBytes());
+            }
+        }
+    }
+
     private void setupResources() {
         Resource resource = new Resource();
-
-        // see MavenProjectBasicStub for details
-        // of getBasedir
 
         // setup default resources
         resource.setDirectory(getBasedir().getPath() + "/src/main/resources");
@@ -86,9 +249,6 @@ public class MavenProjectResourcesStub extends MavenProjectBuildStub {
     private void setupTestResources() {
         Resource resource = new Resource();
 
-        // see MavenProjectBasicStub for details
-        // of getBasedir
-
         // setup default test resources
         resource.setDirectory(getBasedir().getPath() + "/src/test/resources");
         resource.setFiltering(false);
@@ -96,11 +256,42 @@ public class MavenProjectResourcesStub extends MavenProjectBuildStub {
         build.addTestResource(resource);
     }
 
-    public File getBaseDir() {
-        return baseDir == null ? super.getBasedir() : baseDir;
+    @Override
+    public File getBasedir() {
+        return new File(testRootDir);
     }
 
-    public void setBaseDir(File baseDir) {
-        this.baseDir = baseDir;
+    @Override
+    public Build getBuild() {
+        return build;
+    }
+
+    @Override
+    public Properties getProperties() {
+        return properties;
+    }
+
+    @Override
+    public void setDescription(String desc) {
+        description = desc;
+    }
+
+    @Override
+    public String getDescription() {
+        if (description == null) {
+            return "this is a test project";
+        } else {
+            return description;
+        }
+    }
+
+    @Override
+    public List<Resource> getTestResources() {
+        return build.getTestResources();
+    }
+
+    @Override
+    public List<Resource> getResources() {
+        return build.getResources();
     }
 }
