--- src/test/resources/unit/resources-test/plugin-config-relative-path.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Licensed to the Apache Software Foundation (ASF) under one
-  or more contributor license agreements.  See the NOTICE file
-  distributed with this work for additional information
-  regarding copyright ownership.  The ASF licenses this file
-  to you under the Apache License, Version 2.0 (the
-  "License"); you may not use this file except in compliance
-  with the License.  You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing,
-  software distributed under the License is distributed on an
-  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  KIND, either express or implied.  See the License for the
-  specific language governing permissions and limitations
-  under the License.
--->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <name>test3</name>
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>maven-resources-plugin</artifactId>
-       <configuration>
-          <outputDirectory>/test-classes/unit/result-dir</outputDirectory>
-        </configuration>
-      </plugin>
-    </plugins>
-    <resources>
-      <resource>
-        <filtering>false</filtering>
-        <directory>../src/test/resources/unit/test-dir</directory>
-        <includes>
-          **/*.include
-          **/*.test
-        </includes>
-        <excludes>
-          **/*.exclude
-          **/*.nottest
-        </excludes>
-      </resource>
-    </resources>
-  </build>
-</project>
